Keep config code in its own file
